报单机器人已部署

+++ process-bot/src/main/java/com/li/bot/handle/callback/AdminDeleteBusinessCallbackImpl.java
@@ -2,7 +2,10 @@ package com.li.bot.handle.callback;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.li.bot.entity.database.Business;
+import com.li.bot.entity.database.Order;
+import com.li.bot.enums.OrderStatus;
 import com.li.bot.mapper.BusinessMapper;
+import com.li.bot.mapper.OrderMapper;
 import com.li.bot.service.impl.BotServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -10,6 +13,8 @@ import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.methods.updatingmessages.DeleteMessage;
 import org.telegram.telegrambots.meta.api.objects.CallbackQuery;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+import java.util.List;
 
 /**
  * @Author: li
@@ -26,11 +31,22 @@ public class AdminDeleteBusinessCallbackImpl implements ICallback {
     @Autowired
     private BusinessMapper businessMapper;
 
+    @Autowired
+    private OrderMapper orderMapper ;
+
     @Override
     public void execute(BotServiceImpl bot, CallbackQuery callbackQuery) throws TelegramApiException {
 
         String data = callbackQuery.getData();
         String businessId = data.substring(data.lastIndexOf(":") + 1);
+
+
+        List<Order> orderList = orderMapper.getOrderByBusinessId01(Long.valueOf(businessId));
+        if(!orderList.isEmpty()){
+            SendMessage sendMessage = SendMessage.builder().chatId(callbackQuery.getMessage().getChatId()).text("该业务下有未领取订单或处理中的订单，无法删除").build();
+            bot.execute(sendMessage);
+            return;
+        }
 
         int delete = businessMapper.delete(new LambdaQueryWrapper<Business>().eq(Business::getBusinessId, Long.valueOf(businessId)));
         if (delete > 0) {
